# src/main/kotlin/notes/common/pubtator3/pubtator3-openapi.yaml
openapi: 3.0.0
info:
  title: PubTator3 API
  description: |
    This API provides functionality to retrieve biomedical text annotations and related entities.
  version: 1.0.0
servers:
  - url: https://www.ncbi.nlm.nih.gov/research/pubtator3-api/
    description: Main PubTator3 API server

paths:

  /entity/autocomplete/:
    get:
      summary: Find Entity ID by Query and Concept
      description: |
        This endpoint allows users to search for entity identifiers based on a free text query. Optionally, users can specify a bioconcept type and limit the number of returned identifiers.
      parameters:
        - name: query
          in: query
          description: Free text query to search for bioconcept identifiers.
          required: true
          schema:
            type: string
            example: cancer
        - name: concept
          in: query
          description: |
            Optional bioconcept type to filter results. Can be one of the following: `gene`, `disease`, `chemical`, `species`.
          required: false
          schema:
            type: string
            example: disease
        - name: limit
          in: query
          description: |
            Optional parameter to limit the number of results returned. If not specified, the default limit is applied.
          required: false
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: A list of bioconcept identifiers matching or partially matching the query.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: Unique identifier of the bioconcept.
                      example: "@DISEASE_Neoplasms"
                    biotype:
                      type: string
                      description: Type of bioconcept (e.g., disease, gene, chemical).
                      example: disease
                    name:
                      type: string
                      description: Official name of the bioconcept.
                      example: Neoplasms
                    match:
                      type: string
                      description: Information about how the match was made.
                      example: Matched on synonyms <m>Cancer</m>
        '400':
          description: Invalid query or parameters.
        '500':
          description: Server error.

  /search/:
    get:
      summary: Search Publications by Text, Entity ID, or Relations
      description: |
        Query through this API to retrieve relevant publications returned by PubTator3.
        The query can be in the form of free text, entity ID (e.g., @CHEMICAL_remdesivir), or relations between entities.
        Query relations can be specified as relations between two entities (e.g., relations:ANY|@CHEMICAL_Doxorubicin|@DISEASE_Neoplasms) or relations between one entity and entities of a specific type (e.g., relations:ANY|@CHEMICAL_Doxorubicin|DISEASE).
      parameters:
        - name: text
          in: query
          required: true
          description: The free-text query or entity ID to search for.
          schema:
            type: string
          examples:
            example1:
              summary: Search by chemical
              description: Example of searching for a chemical entity
              value: "@CHEMICAL_remdesivir"
            example2:
              summary: Search by relations
              description: Example of searching for relations between entities
              value: "relations:ANY|@CHEMICAL_Doxorubicin|@DISEASE_Neoplasms"
            example3:
              summary: Search by using AND operator
              description: Example of searching for entities using AND operator
              value: "@CHEMICAL_Doxorubicin AND @DISEASE_Neoplasms"
            example4:
              summary: Search free text
              description: Example of searching for free text
              value: "long covid metformin"
        - name: page
          in: query
          required: false
          description: The page number for paginated results (default is 1).
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Successfully retrieved search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: Unique identifier of the publication.
                          example: "37711410"
                        pmid:
                          type: integer
                          description: PubMed ID of the publication.
                          example: 37711410
                        title:
                          type: string
                          description: Title of the publication.
                          example: "Remdesivir."
                        journal:
                          type: string
                          description: Journal where the publication is published.
                          example: "Hosp Pharm"
                        authors:
                          type: array
                          items:
                            type: string
                          example: [ "Levien TL", "Baker DE" ]
                        date:
                          type: string
                          description: Date of publication in ISO 8601 format.
                          example: "2023-10-01T00:00:00Z"
                        doi:
                          type: string
                          description: Digital Object Identifier (DOI) of the publication.
                          example: "10.1177/0018578721999804"
                        score:
                          type: number
                          description: Relevance score for the search result.
                          example: 266.7693
                        text_hl:
                          type: string
                          description: Highlighted search result with matching terms.
                          example: "@<m>CHEMICAL_remdesivir</m> @CHEMICAL_MESH:C000606551 @@@Remdesivir@@@."
                        citations:
                          type: object
                          description: Citation details in different formats.
                          properties:
                            NLM:
                              type: string
                              description: National Library of Medicine citation format.
                              example: "Levien TL, Baker DE. Remdesivir. Hosp Pharm. 2023 Oct;58(5):420-430. PMID: 37711410"
                            BibTeX:
                              type: string
                              description: Citation in BibTeX format.
                              example: "@article{37711410, title={Remdesivir.}, author={Levien TL and Baker DE}, journal={Hosp Pharm}, volume={58}, number={5}, pages={420-430}}"
                  facets:
                    type: object
                    properties:
                      journal:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                              example: "Cureus"
                            type:
                              type: string
                              example: "int"
                            value:
                              type: integer
                              example: 557
                      year:
                        type: array
                        items:
                          type: object
                          properties:
                            name:
                              type: string
                              example: "2023"
                            type:
                              type: string
                              example: "int"
                            value:
                              type: integer
                              example: 3546
                  page_size:
                    type: integer
                    description: Number of results per page.
                    example: 10
                  current:
                    type: integer
                    description: Current page number.
                    example: 1
                  count:
                    type: integer
                    description: Total number of search results.
                    example: 21514
                  total_pages:
                    type: integer
                    description: Total number of pages.
                    example: 2152
        '400':
          description: Invalid request (missing or invalid parameters).
        '404':
          description: No publications found for the query.


  /relations:
    get:
      summary: Find Related Entities
      description: |
        This API allows users to find related entities by providing an `entityId` (found through "Find Entity ID").
        Optionally, users can filter results by `relation_type` and `entity_type`.
      parameters:
        - name: e1
          in: query
          required: true
          description: The entity ID to query related entities for.
          schema:
            type: string
            example: "@CHEMICAL_Doxorubicin"
        - name: type
          in: query
          required: false
          description: The relation type to filter by (optional).
          schema:
            type: string
            enum: [ treat, cause, cotreat, convert, compare, interact, associate, positive_correlate, negative_correlate, prevent, inhibit, stimulate, drug_interact ]
            example: treat
        - name: e2
          in: query
          required: false
          description: The entity type to filter related entities by (optional).
          schema:
            type: string
            enum: [ gene, disease, chemical, variant ]
            example: disease
      responses:
        '200':
          description: Successfully retrieved related entities
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type: string
                      description: Type of relationship (e.g., negative_correlate, treat).
                      example: "negative_correlate"
                    source:
                      type: string
                      description: The source entity ID.
                      example: "@CHEMICAL_ruxolitinib"
                    target:
                      type: string
                      description: The target entity ID.
                      example: "@GENE_JAK1"
                    publications:
                      type: integer
                      description: Number of publications that mention this relationship.
                      example: 571


components:
  schemas:
    Annotation:
      type: object
      properties:
        type:
          type: string
          description: The type of biomedical entity (e.g., gene, chemical, disease).
        start:
          type: integer
          description: The starting index of the annotation in the text.
        end:
          type: integer
          description: The ending index of the annotation in the text.
        text:
          type: string
          description: The annotated text.
        id:
          type: string
          description: The identifier for the entity.